requirements.txt
# CFD Smart Entry System Dependencies

# MetaTrader 5 Integration
MetaTrader5>=5.0.45

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning / AI
scikit-learn>=1.3.0

# Voice Input
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and Utilities
loguru>=0.7.0

# API and Web
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
